many changes
break of single source location development.
Since I had the sources on a stick and took it from one place to the other, the number of files grew and grew because of node_modules directory as adding modules for testing purposes. Even removing those testing modules always left something behind.
Now I want to keep the repository up to date to be able to simply delete development when I want to, or just create a new development environment somewhere. Also I don't need to carry the sources around anymore and I learn which modules I really need to install when time goes by.
So it really seems to be the right thing to do which, by the way, I knew from the start. :-| ;-)